requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
urllib3>=2.0.7 